pyproject.toml
[project]
name = "ask-sg-agencies"
version = "0.1.0"
description = "SG Open Data Agentic Dataset Recommender"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "streamlit>=1.38.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
